fix: verify credential before storing
This one is a bit tricky. Whenever a credential that is revocable is
being verified, the dependencies we use for verification require a
function that checks the revocation status. However, we currently do not
suppport revocation for W3C formatted credentials (`RevocationList2020`).

Instead of throwing an error whenever we receive a revocable credential,
I added a boolean flag to the function calls to specify if the
revocation status should be verified, which is set to `true` by default.

If a revocable credential is received and the user explicitly sets this
flag to `false`, a function is passed that will pretend like the
revocation status is valid. On the other hand, if a revocable credential
is received and this flag is set to `true`, an error is thrown
indicating that revocation isn't supported for the W3C format.

This workaround allows us to receive, verify, and store revocable
credentials, regardless of AFJ's support for `RevocationList2020`. The
drawback of this approach is that a holder has no way of checking if a
credential has been revoked or not, before sending it off to a verifier.

Signed-off-by: Karim Stekelenburg <karim@animo.id>